+++ .github/workflows/pull_request.yaml
@@ -36,7 +36,7 @@ jobs:
       - name: Lint with Clippy
         run: cargo clippy
       - name: Test with Cargo
-        run: cargo test
+        run: cargo test --workspace
   golang:
     runs-on: ubuntu-latest
     steps:
