+++ Cargo.toml
@@ -8,17 +8,10 @@ authors = [
 ]
 
 [workspace]
-members = ["crate-proto"]
+members = ["crate-proto", "crate-utils", "services/*"]
 
 [workspace.dependencies]
 prost = "0.13.4"
-
-[dependencies]
-axum = { version = "0.7.9", features = ["multipart"] }
-dotenv = "0.15.0"
-tokio = { version = "1.41.1", features = ["full"] }
-garde = { version = "0.20.0", features = ["derive", "serde"] }
-hyper-util = "0.1.10"
 serde = { version = "1.0.215", features = ["derive"] }
 sqlx = { version = "0.8.2", features = [
   "runtime-tokio",
@@ -31,10 +24,19 @@ sqlx = { version = "0.8.2", features = [
   "chrono",
   "json",
 ] }
-tokio-stream = "0.1.17"
 tonic = "0.12.3"
-tonic-web = "0.12.3"
+jwt = "0.16.0"
+tokio = { version = "1.41.1", features = ["full"] }
+tracing = "0.1.41"
+tracing-subscriber = { version = "0.3.19", features = ["env-filter"] }
+tokio-stream = "0.1.17"
 tower = "0.5.2"
+futures = "0.3.31"
+garde = { version = "0.20.0", features = ["derive", "serde"] }
+uuid = { version = "1.12.1", features = ["v4"] }
+
+[dependencies]
+dotenv = "0.15.0"
 tower-http = { version = "0.6.2", features = [
   "catch-panic",
   "compression-gzip",
@@ -47,20 +49,10 @@ tower-http = { version = "0.6.2", features = [
   "timeout",
   "trace",
 ] }
-tracing = "0.1.41"
-tracing-subscriber = { version = "0.3.19", features = ["env-filter"] }
-sea-orm = { version = "1.1.3", features = [
-  "sqlx-postgres",
-  "runtime-tokio-rustls",
-] }
-futures = "0.3.31"
 test-log = "0.2.16"
-jwt = "0.16.0"
-sha2 = "0.10.8"
-hmac = "0.12.1"
 tokio-util = "0.7.13"
 clap = { version = "4.5.26", features = ["derive", "string"] }
-tempdir = "0.3.7"
+
 
 [build-dependencies]
 tonic-build = "0.12.3"
