+++ crate-proto/Cargo.toml
@@ -5,3 +5,4 @@ edition = "2024"
 
 [dependencies]
 prost = { workspace = true }
+tonic = { workspace = true }
