--- crate-proto/src/google.protobuf.rs
@@ -1,11 +1,0 @@
-// This file is @generated by prost-build.
-/// A generic empty message that you can re-use to avoid defining duplicated
-/// empty messages in your APIs. A typical example is to use it as the request
-/// or the response type of an API method. For instance:
-///
-///      service Foo {
-///        rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty);
-///      }
-///
-#[derive(Clone, Copy, PartialEq, ::prost::Message)]
-pub struct Empty {}
