+++ crate-utils/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "crate-utils"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+hyper-util = "0.1.10"
+tonic = { workspace = true }
+tokio = { workspace = true }
+tokio-stream = { workspace = true }
+tower = { workspace = true }
