+++ services/authentication/Cargo.toml
@@ -4,3 +4,9 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+crate-proto = { path = "../../crate-proto" }
+crate-utils = { path = "../../crate-utils" }
+sqlx = { workspace = true }
+serde = { workspace = true }
+tonic = { workspace = true }
+tracing = { workspace = true }
