+++ services/storage/Cargo.toml
@@ -4,3 +4,13 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+sqlx = { workspace = true }
+tokio = { workspace = true }
+tonic = { workspace = true }
+tokio-stream = { workspace = true }
+crate-proto = { path = "../../crate-proto" }
+crate-utils = { path = "../../crate-utils" }
+futures = { workspace = true }
+tempdir = "0.3.7"
+uuid = { workspace = true }
+tracing = { workspace = true }
