--- src/controllers/mod.rs
@@ -1,4 +1,0 @@
-pub mod auth;
-pub mod employee;
-pub mod storage;
-pub mod user;
