--- src/models/_entities/department.rs
@@ -1,36 +1,0 @@
-//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.3
-
-use sea_orm::entity::prelude::*;
-
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
-#[sea_orm(schema_name = "etmar_logistics", table_name = "department")]
-pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub id: Uuid,
-    pub name: String,
-    #[sea_orm(column_type = "Text", nullable)]
-    pub description: Option<String>,
-    pub manager_id: Uuid,
-    pub created_at: DateTime,
-    pub updated_at: DateTime,
-}
-
-#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
-pub enum Relation {
-    #[sea_orm(
-        belongs_to = "super::employee::Entity",
-        from = "Column::ManagerId",
-        to = "super::employee::Column::Id",
-        on_update = "NoAction",
-        on_delete = "NoAction"
-    )]
-    Employee,
-}
-
-impl Related<super::employee::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::Employee.def()
-    }
-}
-
-impl ActiveModelBehavior for ActiveModel {}
