--- src/models/_entities/employee.rs
@@ -1,76 +1,0 @@
-//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.3
-
-use sea_orm::entity::prelude::*;
-
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
-#[sea_orm(schema_name = "etmar_logistics", table_name = "employee")]
-pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub id: Uuid,
-    pub user_id: Uuid,
-    pub first_name: String,
-    pub middle_name: Option<String>,
-    pub last_name: String,
-    pub sex: String,
-    #[sea_orm(column_type = "Text")]
-    pub address: String,
-    #[sea_orm(column_type = "Text")]
-    pub position: String,
-    #[sea_orm(column_type = "Text")]
-    pub contact_number: String,
-    pub contract_type: String,
-    pub birth_day: Date,
-    pub created_at: DateTime,
-    pub updated_at: DateTime,
-    pub department_id: Uuid,
-}
-
-#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
-pub enum Relation {
-    #[sea_orm(
-        belongs_to = "super::department::Entity",
-        from = "Column::DepartmentId",
-        to = "super::department::Column::Id",
-        on_update = "NoAction",
-        on_delete = "NoAction"
-    )]
-    Department,
-    #[sea_orm(has_many = "super::task::Entity")]
-    Task,
-    #[sea_orm(has_many = "super::task_comment::Entity")]
-    TaskComment,
-    #[sea_orm(
-        belongs_to = "super::user::Entity",
-        from = "Column::UserId",
-        to = "super::user::Column::Id",
-        on_update = "NoAction",
-        on_delete = "NoAction"
-    )]
-    User,
-}
-
-impl Related<super::department::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::Department.def()
-    }
-}
-
-impl Related<super::task::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::Task.def()
-    }
-}
-
-impl Related<super::task_comment::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::TaskComment.def()
-    }
-}
-
-impl Related<super::user::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::User.def()
-    }
-}
-
-impl ActiveModelBehavior for ActiveModel {}
