--- src/models/_entities/file.rs
@@ -1,45 +1,0 @@
-//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.3
-
-use sea_orm::entity::prelude::*;
-
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
-#[sea_orm(schema_name = "etmar_logistics", table_name = "file")]
-pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub id: Uuid,
-    #[sea_orm(unique)]
-    pub name: String,
-    #[sea_orm(column_type = "Text")]
-    pub path: String,
-    pub owner_id: Uuid,
-    pub created_at: DateTime,
-    pub updated_at: DateTime,
-}
-
-#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
-pub enum Relation {
-    #[sea_orm(has_many = "super::task_comment_file_attachment::Entity")]
-    TaskCommentFileAttachment,
-    #[sea_orm(
-        belongs_to = "super::user::Entity",
-        from = "Column::OwnerId",
-        to = "super::user::Column::Id",
-        on_update = "NoAction",
-        on_delete = "NoAction"
-    )]
-    User,
-}
-
-impl Related<super::task_comment_file_attachment::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::TaskCommentFileAttachment.def()
-    }
-}
-
-impl Related<super::user::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::User.def()
-    }
-}
-
-impl ActiveModelBehavior for ActiveModel {}
