--- src/models/_entities/mod.rs
@@ -1,11 +1,0 @@
-//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.3
-
-pub mod prelude;
-
-pub mod department;
-pub mod employee;
-pub mod file;
-pub mod task;
-pub mod task_comment;
-pub mod task_comment_file_attachment;
-pub mod user;
