--- src/models/_entities/prelude.rs
@@ -1,9 +1,0 @@
-//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.3
-
-pub use super::department::Entity as Department;
-pub use super::employee::Entity as Employee;
-pub use super::file::Entity as File;
-pub use super::task::Entity as Task;
-pub use super::task_comment::Entity as TaskComment;
-pub use super::task_comment_file_attachment::Entity as TaskCommentFileAttachment;
-pub use super::user::Entity as User;
