--- src/models/_entities/sea_orm_active_enums.rs
@@ -1,48 +1,0 @@
-//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.3
-
-use sea_orm::entity::prelude::*;
-
-#[derive(Debug, Clone, PartialEq, Eq, EnumIter, DeriveActiveEnum)]
-#[sea_orm(rs_type = "String", db_type = "Enum", enum_name = "contract_type_enum")]
-pub enum ContractTypeEnum {
-    #[sea_orm(string_value = "FULL_TIME")]
-    FullTime,
-    #[sea_orm(string_value = "PART_TIME")]
-    PartTime,
-}
-#[derive(Debug, Clone, PartialEq, Eq, EnumIter, DeriveActiveEnum)]
-#[sea_orm(
-    rs_type = "String",
-    db_type = "Enum",
-    enum_name = "etmar_logistics.role_enum"
-)]
-pub enum RoleEnum {
-    #[sea_orm(string_value = "SUPER_ADMIN")]
-    SuperAdmin,
-    #[sea_orm(string_value = "ADMIN")]
-    Admin,
-    #[sea_orm(string_value = "EMPLOYEE")]
-    Employee,
-    #[sea_orm(string_value = "CLIENT")]
-    Client,
-}
-#[derive(Debug, Clone, PartialEq, Eq, EnumIter, DeriveActiveEnum)]
-#[sea_orm(rs_type = "String", db_type = "Enum", enum_name = "sex_enum")]
-pub enum SexEnum {
-    #[sea_orm(string_value = "MALE")]
-    Male,
-    #[sea_orm(string_value = "FEMALE")]
-    Female,
-}
-#[derive(Debug, Clone, PartialEq, Eq, EnumIter, DeriveActiveEnum)]
-#[sea_orm(rs_type = "String", db_type = "Enum", enum_name = "task_status_enum")]
-pub enum TaskStatusEnum {
-    #[sea_orm(string_value = "UNASSIGNED")]
-    Unassigned,
-    #[sea_orm(string_value = "IN_PROGRESS")]
-    InProgress,
-    #[sea_orm(string_value = "REVIEW")]
-    Review,
-    #[sea_orm(string_value = "DONE")]
-    Done,
-}
