--- src/models/_entities/task.rs
@@ -1,46 +1,0 @@
-//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.3
-
-use sea_orm::entity::prelude::*;
-
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
-#[sea_orm(schema_name = "etmar_logistics", table_name = "task")]
-pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub id: Uuid,
-    pub title: String,
-    #[sea_orm(column_type = "Text", nullable)]
-    pub description: Option<String>,
-    pub status: String,
-    pub issued_by_employee_id: Uuid,
-    pub deadline: Date,
-    pub created_at: DateTime,
-    pub updated_at: DateTime,
-}
-
-#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
-pub enum Relation {
-    #[sea_orm(
-        belongs_to = "super::employee::Entity",
-        from = "Column::IssuedByEmployeeId",
-        to = "super::employee::Column::Id",
-        on_update = "NoAction",
-        on_delete = "NoAction"
-    )]
-    Employee,
-    #[sea_orm(has_many = "super::task_comment::Entity")]
-    TaskComment,
-}
-
-impl Related<super::employee::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::Employee.def()
-    }
-}
-
-impl Related<super::task_comment::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::TaskComment.def()
-    }
-}
-
-impl ActiveModelBehavior for ActiveModel {}
