--- src/models/_entities/user.rs
@@ -1,39 +1,0 @@
-//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.3
-
-use sea_orm::entity::prelude::*;
-
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
-#[sea_orm(schema_name = "etmar_logistics", table_name = "user")]
-pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub id: Uuid,
-    #[sea_orm(unique)]
-    pub email: String,
-    #[sea_orm(unique)]
-    pub password: String,
-    pub user_role: String,
-    pub created_at: DateTime,
-    pub updated_at: DateTime,
-}
-
-#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
-pub enum Relation {
-    #[sea_orm(has_many = "super::employee::Entity")]
-    Employee,
-    #[sea_orm(has_many = "super::file::Entity")]
-    File,
-}
-
-impl Related<super::employee::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::Employee.def()
-    }
-}
-
-impl Related<super::file::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::File.def()
-    }
-}
-
-impl ActiveModelBehavior for ActiveModel {}
