--- src/models/_proto/file_management.rs
@@ -1,1094 +1,0 @@
-// This file is @generated by prost-build.
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct CreateFileRequest {
-    #[prost(string, tag = "1")]
-    pub name: ::prost::alloc::string::String,
-    #[prost(string, tag = "2")]
-    pub path: ::prost::alloc::string::String,
-    #[prost(string, tag = "3")]
-    pub mime_type: ::prost::alloc::string::String,
-    #[prost(uint64, tag = "4")]
-    pub size: u64,
-    #[prost(enumeration = "FileStatus", tag = "5")]
-    pub status: i32,
-    #[prost(string, tag = "6")]
-    pub system_origin: ::prost::alloc::string::String,
-    #[prost(string, tag = "7")]
-    pub created_by: ::prost::alloc::string::String,
-    #[prost(string, repeated, tag = "8")]
-    pub tags: ::prost::alloc::vec::Vec<::prost::alloc::string::String>,
-    #[prost(string, tag = "9")]
-    pub description: ::prost::alloc::string::String,
-}
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct UploadFileRequest {
-    #[prost(oneof = "upload_file_request::Data", tags = "1, 2")]
-    pub data: ::core::option::Option<upload_file_request::Data>,
-}
-/// Nested message and enum types in `UploadFileRequest`.
-pub mod upload_file_request {
-    #[derive(Clone, PartialEq, ::prost::Oneof)]
-    pub enum Data {
-        #[prost(message, tag = "1")]
-        FileInfo(super::FileInfo),
-        #[prost(bytes, tag = "2")]
-        ChunkData(::prost::alloc::vec::Vec<u8>),
-    }
-}
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct FileInfo {
-    #[prost(string, tag = "1")]
-    pub name: ::prost::alloc::string::String,
-    #[prost(string, tag = "2")]
-    pub mime_type: ::prost::alloc::string::String,
-    #[prost(string, tag = "3")]
-    pub system_origin: ::prost::alloc::string::String,
-    #[prost(string, tag = "4")]
-    pub created_by: ::prost::alloc::string::String,
-    #[prost(string, repeated, tag = "5")]
-    pub tags: ::prost::alloc::vec::Vec<::prost::alloc::string::String>,
-    #[prost(string, tag = "6")]
-    pub description: ::prost::alloc::string::String,
-}
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct GetFileRequest {
-    #[prost(string, tag = "1")]
-    pub file_id: ::prost::alloc::string::String,
-}
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct ListFilesRequest {
-    #[prost(uint32, tag = "1")]
-    pub page: u32,
-    #[prost(uint32, tag = "2")]
-    pub page_size: u32,
-    #[prost(enumeration = "FileStatus", tag = "3")]
-    pub status: i32,
-    #[prost(string, tag = "4")]
-    pub system_origin: ::prost::alloc::string::String,
-    #[prost(string, repeated, tag = "5")]
-    pub tags: ::prost::alloc::vec::Vec<::prost::alloc::string::String>,
-    #[prost(string, tag = "6")]
-    pub search_term: ::prost::alloc::string::String,
-    #[prost(string, tag = "7")]
-    pub created_by: ::prost::alloc::string::String,
-    #[prost(enumeration = "FileSort", tag = "8")]
-    pub sort_by: i32,
-    #[prost(bool, tag = "9")]
-    pub ascending: bool,
-}
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct UpdateFileRequest {
-    #[prost(string, tag = "1")]
-    pub file_id: ::prost::alloc::string::String,
-    #[prost(string, optional, tag = "2")]
-    pub name: ::core::option::Option<::prost::alloc::string::String>,
-    #[prost(string, optional, tag = "3")]
-    pub path: ::core::option::Option<::prost::alloc::string::String>,
-    #[prost(string, optional, tag = "4")]
-    pub description: ::core::option::Option<::prost::alloc::string::String>,
-    #[prost(string, repeated, tag = "5")]
-    pub tags: ::prost::alloc::vec::Vec<::prost::alloc::string::String>,
-}
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct UpdateFileStatusRequest {
-    #[prost(string, tag = "1")]
-    pub file_id: ::prost::alloc::string::String,
-    #[prost(enumeration = "FileStatus", tag = "2")]
-    pub status: i32,
-}
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct DeleteFileRequest {
-    #[prost(string, tag = "1")]
-    pub file_id: ::prost::alloc::string::String,
-    #[prost(bool, tag = "2")]
-    pub hard_delete: bool,
-}
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct DownloadFileRequest {
-    #[prost(string, tag = "1")]
-    pub file_id: ::prost::alloc::string::String,
-}
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct FileResponse {
-    #[prost(string, tag = "1")]
-    pub id: ::prost::alloc::string::String,
-    #[prost(string, tag = "2")]
-    pub name: ::prost::alloc::string::String,
-    #[prost(string, tag = "3")]
-    pub path: ::prost::alloc::string::String,
-    #[prost(string, tag = "4")]
-    pub mime_type: ::prost::alloc::string::String,
-    #[prost(uint64, tag = "5")]
-    pub size: u64,
-    #[prost(enumeration = "FileStatus", tag = "6")]
-    pub status: i32,
-    #[prost(string, tag = "7")]
-    pub system_origin: ::prost::alloc::string::String,
-    #[prost(string, tag = "8")]
-    pub created_by: ::prost::alloc::string::String,
-    #[prost(string, tag = "9")]
-    pub created_at: ::prost::alloc::string::String,
-    #[prost(string, tag = "10")]
-    pub updated_at: ::prost::alloc::string::String,
-    #[prost(string, repeated, tag = "11")]
-    pub tags: ::prost::alloc::vec::Vec<::prost::alloc::string::String>,
-    #[prost(string, tag = "12")]
-    pub description: ::prost::alloc::string::String,
-    #[prost(string, optional, tag = "13")]
-    pub last_accessed_at: ::core::option::Option<::prost::alloc::string::String>,
-    #[prost(string, optional, tag = "14")]
-    pub deleted_at: ::core::option::Option<::prost::alloc::string::String>,
-}
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct ListFilesResponse {
-    #[prost(message, repeated, tag = "1")]
-    pub files: ::prost::alloc::vec::Vec<FileResponse>,
-    #[prost(uint32, tag = "2")]
-    pub total_count: u32,
-    #[prost(uint32, tag = "3")]
-    pub page: u32,
-    #[prost(uint32, tag = "4")]
-    pub page_size: u32,
-}
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct DeleteFileResponse {
-    #[prost(string, tag = "1")]
-    pub file_id: ::prost::alloc::string::String,
-    #[prost(bool, tag = "2")]
-    pub success: bool,
-}
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct FileChunk {
-    #[prost(bytes = "vec", tag = "1")]
-    pub content: ::prost::alloc::vec::Vec<u8>,
-}
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct FileError {
-    #[prost(string, tag = "1")]
-    pub file_id: ::prost::alloc::string::String,
-    #[prost(string, tag = "2")]
-    pub error_message: ::prost::alloc::string::String,
-    #[prost(enumeration = "FileErrorCode", tag = "3")]
-    pub error_code: i32,
-}
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration)]
-#[repr(i32)]
-pub enum FileStatus {
-    Unspecified = 0,
-    Active = 1,
-    Archived = 2,
-    Deleted = 3,
-}
-impl FileStatus {
-    /// String value of the enum field names used in the ProtoBuf definition.
-    ///
-    /// The values are not transformed in any way and thus are considered stable
-    /// (if the ProtoBuf definition does not change) and safe for programmatic use.
-    pub fn as_str_name(&self) -> &'static str {
-        match self {
-            Self::Unspecified => "FILE_STATUS_UNSPECIFIED",
-            Self::Active => "FILE_STATUS_ACTIVE",
-            Self::Archived => "FILE_STATUS_ARCHIVED",
-            Self::Deleted => "FILE_STATUS_DELETED",
-        }
-    }
-    /// Creates an enum from field names used in the ProtoBuf definition.
-    pub fn from_str_name(value: &str) -> ::core::option::Option<Self> {
-        match value {
-            "FILE_STATUS_UNSPECIFIED" => Some(Self::Unspecified),
-            "FILE_STATUS_ACTIVE" => Some(Self::Active),
-            "FILE_STATUS_ARCHIVED" => Some(Self::Archived),
-            "FILE_STATUS_DELETED" => Some(Self::Deleted),
-            _ => None,
-        }
-    }
-}
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration)]
-#[repr(i32)]
-pub enum FileSort {
-    Unspecified = 0,
-    Name = 1,
-    CreatedAt = 2,
-    UpdatedAt = 3,
-    Size = 4,
-}
-impl FileSort {
-    /// String value of the enum field names used in the ProtoBuf definition.
-    ///
-    /// The values are not transformed in any way and thus are considered stable
-    /// (if the ProtoBuf definition does not change) and safe for programmatic use.
-    pub fn as_str_name(&self) -> &'static str {
-        match self {
-            Self::Unspecified => "FILE_SORT_UNSPECIFIED",
-            Self::Name => "FILE_SORT_NAME",
-            Self::CreatedAt => "FILE_SORT_CREATED_AT",
-            Self::UpdatedAt => "FILE_SORT_UPDATED_AT",
-            Self::Size => "FILE_SORT_SIZE",
-        }
-    }
-    /// Creates an enum from field names used in the ProtoBuf definition.
-    pub fn from_str_name(value: &str) -> ::core::option::Option<Self> {
-        match value {
-            "FILE_SORT_UNSPECIFIED" => Some(Self::Unspecified),
-            "FILE_SORT_NAME" => Some(Self::Name),
-            "FILE_SORT_CREATED_AT" => Some(Self::CreatedAt),
-            "FILE_SORT_UPDATED_AT" => Some(Self::UpdatedAt),
-            "FILE_SORT_SIZE" => Some(Self::Size),
-            _ => None,
-        }
-    }
-}
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration)]
-#[repr(i32)]
-pub enum FileErrorCode {
-    FileErrorUnspecified = 0,
-    FileErrorNotFound = 1,
-    FileErrorPermissionDenied = 2,
-    FileErrorInvalidStatus = 3,
-    FileErrorStorageFailed = 4,
-    FileErrorSizeExceeded = 5,
-    FileErrorInvalidType = 6,
-}
-impl FileErrorCode {
-    /// String value of the enum field names used in the ProtoBuf definition.
-    ///
-    /// The values are not transformed in any way and thus are considered stable
-    /// (if the ProtoBuf definition does not change) and safe for programmatic use.
-    pub fn as_str_name(&self) -> &'static str {
-        match self {
-            Self::FileErrorUnspecified => "FILE_ERROR_UNSPECIFIED",
-            Self::FileErrorNotFound => "FILE_ERROR_NOT_FOUND",
-            Self::FileErrorPermissionDenied => "FILE_ERROR_PERMISSION_DENIED",
-            Self::FileErrorInvalidStatus => "FILE_ERROR_INVALID_STATUS",
-            Self::FileErrorStorageFailed => "FILE_ERROR_STORAGE_FAILED",
-            Self::FileErrorSizeExceeded => "FILE_ERROR_SIZE_EXCEEDED",
-            Self::FileErrorInvalidType => "FILE_ERROR_INVALID_TYPE",
-        }
-    }
-    /// Creates an enum from field names used in the ProtoBuf definition.
-    pub fn from_str_name(value: &str) -> ::core::option::Option<Self> {
-        match value {
-            "FILE_ERROR_UNSPECIFIED" => Some(Self::FileErrorUnspecified),
-            "FILE_ERROR_NOT_FOUND" => Some(Self::FileErrorNotFound),
-            "FILE_ERROR_PERMISSION_DENIED" => Some(Self::FileErrorPermissionDenied),
-            "FILE_ERROR_INVALID_STATUS" => Some(Self::FileErrorInvalidStatus),
-            "FILE_ERROR_STORAGE_FAILED" => Some(Self::FileErrorStorageFailed),
-            "FILE_ERROR_SIZE_EXCEEDED" => Some(Self::FileErrorSizeExceeded),
-            "FILE_ERROR_INVALID_TYPE" => Some(Self::FileErrorInvalidType),
-            _ => None,
-        }
-    }
-}
-/// Generated client implementations.
-pub mod file_management_client {
-    #![allow(
-        unused_variables,
-        dead_code,
-        missing_docs,
-        clippy::wildcard_imports,
-        clippy::let_unit_value,
-    )]
-    use tonic::codegen::*;
-    use tonic::codegen::http::Uri;
-    #[derive(Debug, Clone)]
-    pub struct FileManagementClient<T> {
-        inner: tonic::client::Grpc<T>,
-    }
-    impl FileManagementClient<tonic::transport::Channel> {
-        /// Attempt to create a new client by connecting to a given endpoint.
-        pub async fn connect<D>(dst: D) -> Result<Self, tonic::transport::Error>
-        where
-            D: TryInto<tonic::transport::Endpoint>,
-            D::Error: Into<StdError>,
-        {
-            let conn = tonic::transport::Endpoint::new(dst)?.connect().await?;
-            Ok(Self::new(conn))
-        }
-    }
-    impl<T> FileManagementClient<T>
-    where
-        T: tonic::client::GrpcService<tonic::body::BoxBody>,
-        T::Error: Into<StdError>,
-        T::ResponseBody: Body<Data = Bytes> + std::marker::Send + 'static,
-        <T::ResponseBody as Body>::Error: Into<StdError> + std::marker::Send,
-    {
-        pub fn new(inner: T) -> Self {
-            let inner = tonic::client::Grpc::new(inner);
-            Self { inner }
-        }
-        pub fn with_origin(inner: T, origin: Uri) -> Self {
-            let inner = tonic::client::Grpc::with_origin(inner, origin);
-            Self { inner }
-        }
-        pub fn with_interceptor<F>(
-            inner: T,
-            interceptor: F,
-        ) -> FileManagementClient<InterceptedService<T, F>>
-        where
-            F: tonic::service::Interceptor,
-            T::ResponseBody: Default,
-            T: tonic::codegen::Service<
-                http::Request<tonic::body::BoxBody>,
-                Response = http::Response<
-                    <T as tonic::client::GrpcService<tonic::body::BoxBody>>::ResponseBody,
-                >,
-            >,
-            <T as tonic::codegen::Service<
-                http::Request<tonic::body::BoxBody>,
-            >>::Error: Into<StdError> + std::marker::Send + std::marker::Sync,
-        {
-            FileManagementClient::new(InterceptedService::new(inner, interceptor))
-        }
-        /// Compress requests with the given encoding.
-        ///
-        /// This requires the server to support it otherwise it might respond with an
-        /// error.
-        #[must_use]
-        pub fn send_compressed(mut self, encoding: CompressionEncoding) -> Self {
-            self.inner = self.inner.send_compressed(encoding);
-            self
-        }
-        /// Enable decompressing responses.
-        #[must_use]
-        pub fn accept_compressed(mut self, encoding: CompressionEncoding) -> Self {
-            self.inner = self.inner.accept_compressed(encoding);
-            self
-        }
-        /// Limits the maximum size of a decoded message.
-        ///
-        /// Default: `4MB`
-        #[must_use]
-        pub fn max_decoding_message_size(mut self, limit: usize) -> Self {
-            self.inner = self.inner.max_decoding_message_size(limit);
-            self
-        }
-        /// Limits the maximum size of an encoded message.
-        ///
-        /// Default: `usize::MAX`
-        #[must_use]
-        pub fn max_encoding_message_size(mut self, limit: usize) -> Self {
-            self.inner = self.inner.max_encoding_message_size(limit);
-            self
-        }
-        pub async fn create_file(
-            &mut self,
-            request: impl tonic::IntoRequest<super::CreateFileRequest>,
-        ) -> std::result::Result<tonic::Response<super::FileResponse>, tonic::Status> {
-            self.inner
-                .ready()
-                .await
-                .map_err(|e| {
-                    tonic::Status::unknown(
-                        format!("Service was not ready: {}", e.into()),
-                    )
-                })?;
-            let codec = tonic::codec::ProstCodec::default();
-            let path = http::uri::PathAndQuery::from_static(
-                "/file_management.FileManagement/CreateFile",
-            );
-            let mut req = request.into_request();
-            req.extensions_mut()
-                .insert(GrpcMethod::new("file_management.FileManagement", "CreateFile"));
-            self.inner.unary(req, path, codec).await
-        }
-        pub async fn upload_file(
-            &mut self,
-            request: impl tonic::IntoStreamingRequest<Message = super::UploadFileRequest>,
-        ) -> std::result::Result<tonic::Response<super::FileResponse>, tonic::Status> {
-            self.inner
-                .ready()
-                .await
-                .map_err(|e| {
-                    tonic::Status::unknown(
-                        format!("Service was not ready: {}", e.into()),
-                    )
-                })?;
-            let codec = tonic::codec::ProstCodec::default();
-            let path = http::uri::PathAndQuery::from_static(
-                "/file_management.FileManagement/UploadFile",
-            );
-            let mut req = request.into_streaming_request();
-            req.extensions_mut()
-                .insert(GrpcMethod::new("file_management.FileManagement", "UploadFile"));
-            self.inner.client_streaming(req, path, codec).await
-        }
-        pub async fn get_file(
-            &mut self,
-            request: impl tonic::IntoRequest<super::GetFileRequest>,
-        ) -> std::result::Result<tonic::Response<super::FileResponse>, tonic::Status> {
-            self.inner
-                .ready()
-                .await
-                .map_err(|e| {
-                    tonic::Status::unknown(
-                        format!("Service was not ready: {}", e.into()),
-                    )
-                })?;
-            let codec = tonic::codec::ProstCodec::default();
-            let path = http::uri::PathAndQuery::from_static(
-                "/file_management.FileManagement/GetFile",
-            );
-            let mut req = request.into_request();
-            req.extensions_mut()
-                .insert(GrpcMethod::new("file_management.FileManagement", "GetFile"));
-            self.inner.unary(req, path, codec).await
-        }
-        pub async fn list_files(
-            &mut self,
-            request: impl tonic::IntoRequest<super::ListFilesRequest>,
-        ) -> std::result::Result<
-            tonic::Response<super::ListFilesResponse>,
-            tonic::Status,
-        > {
-            self.inner
-                .ready()
-                .await
-                .map_err(|e| {
-                    tonic::Status::unknown(
-                        format!("Service was not ready: {}", e.into()),
-                    )
-                })?;
-            let codec = tonic::codec::ProstCodec::default();
-            let path = http::uri::PathAndQuery::from_static(
-                "/file_management.FileManagement/ListFiles",
-            );
-            let mut req = request.into_request();
-            req.extensions_mut()
-                .insert(GrpcMethod::new("file_management.FileManagement", "ListFiles"));
-            self.inner.unary(req, path, codec).await
-        }
-        pub async fn download_file(
-            &mut self,
-            request: impl tonic::IntoRequest<super::DownloadFileRequest>,
-        ) -> std::result::Result<
-            tonic::Response<tonic::codec::Streaming<super::FileChunk>>,
-            tonic::Status,
-        > {
-            self.inner
-                .ready()
-                .await
-                .map_err(|e| {
-                    tonic::Status::unknown(
-                        format!("Service was not ready: {}", e.into()),
-                    )
-                })?;
-            let codec = tonic::codec::ProstCodec::default();
-            let path = http::uri::PathAndQuery::from_static(
-                "/file_management.FileManagement/DownloadFile",
-            );
-            let mut req = request.into_request();
-            req.extensions_mut()
-                .insert(
-                    GrpcMethod::new("file_management.FileManagement", "DownloadFile"),
-                );
-            self.inner.server_streaming(req, path, codec).await
-        }
-        pub async fn update_file(
-            &mut self,
-            request: impl tonic::IntoRequest<super::UpdateFileRequest>,
-        ) -> std::result::Result<tonic::Response<super::FileResponse>, tonic::Status> {
-            self.inner
-                .ready()
-                .await
-                .map_err(|e| {
-                    tonic::Status::unknown(
-                        format!("Service was not ready: {}", e.into()),
-                    )
-                })?;
-            let codec = tonic::codec::ProstCodec::default();
-            let path = http::uri::PathAndQuery::from_static(
-                "/file_management.FileManagement/UpdateFile",
-            );
-            let mut req = request.into_request();
-            req.extensions_mut()
-                .insert(GrpcMethod::new("file_management.FileManagement", "UpdateFile"));
-            self.inner.unary(req, path, codec).await
-        }
-        pub async fn update_file_status(
-            &mut self,
-            request: impl tonic::IntoRequest<super::UpdateFileStatusRequest>,
-        ) -> std::result::Result<tonic::Response<super::FileResponse>, tonic::Status> {
-            self.inner
-                .ready()
-                .await
-                .map_err(|e| {
-                    tonic::Status::unknown(
-                        format!("Service was not ready: {}", e.into()),
-                    )
-                })?;
-            let codec = tonic::codec::ProstCodec::default();
-            let path = http::uri::PathAndQuery::from_static(
-                "/file_management.FileManagement/UpdateFileStatus",
-            );
-            let mut req = request.into_request();
-            req.extensions_mut()
-                .insert(
-                    GrpcMethod::new("file_management.FileManagement", "UpdateFileStatus"),
-                );
-            self.inner.unary(req, path, codec).await
-        }
-        pub async fn delete_file(
-            &mut self,
-            request: impl tonic::IntoRequest<super::DeleteFileRequest>,
-        ) -> std::result::Result<
-            tonic::Response<super::DeleteFileResponse>,
-            tonic::Status,
-        > {
-            self.inner
-                .ready()
-                .await
-                .map_err(|e| {
-                    tonic::Status::unknown(
-                        format!("Service was not ready: {}", e.into()),
-                    )
-                })?;
-            let codec = tonic::codec::ProstCodec::default();
-            let path = http::uri::PathAndQuery::from_static(
-                "/file_management.FileManagement/DeleteFile",
-            );
-            let mut req = request.into_request();
-            req.extensions_mut()
-                .insert(GrpcMethod::new("file_management.FileManagement", "DeleteFile"));
-            self.inner.unary(req, path, codec).await
-        }
-    }
-}
-/// Generated server implementations.
-pub mod file_management_server {
-    #![allow(
-        unused_variables,
-        dead_code,
-        missing_docs,
-        clippy::wildcard_imports,
-        clippy::let_unit_value,
-    )]
-    use tonic::codegen::*;
-    /// Generated trait containing gRPC methods that should be implemented for use with FileManagementServer.
-    #[async_trait]
-    pub trait FileManagement: std::marker::Send + std::marker::Sync + 'static {
-        async fn create_file(
-            &self,
-            request: tonic::Request<super::CreateFileRequest>,
-        ) -> std::result::Result<tonic::Response<super::FileResponse>, tonic::Status>;
-        async fn upload_file(
-            &self,
-            request: tonic::Request<tonic::Streaming<super::UploadFileRequest>>,
-        ) -> std::result::Result<tonic::Response<super::FileResponse>, tonic::Status>;
-        async fn get_file(
-            &self,
-            request: tonic::Request<super::GetFileRequest>,
-        ) -> std::result::Result<tonic::Response<super::FileResponse>, tonic::Status>;
-        async fn list_files(
-            &self,
-            request: tonic::Request<super::ListFilesRequest>,
-        ) -> std::result::Result<
-            tonic::Response<super::ListFilesResponse>,
-            tonic::Status,
-        >;
-        /// Server streaming response type for the DownloadFile method.
-        type DownloadFileStream: tonic::codegen::tokio_stream::Stream<
-                Item = std::result::Result<super::FileChunk, tonic::Status>,
-            >
-            + std::marker::Send
-            + 'static;
-        async fn download_file(
-            &self,
-            request: tonic::Request<super::DownloadFileRequest>,
-        ) -> std::result::Result<
-            tonic::Response<Self::DownloadFileStream>,
-            tonic::Status,
-        >;
-        async fn update_file(
-            &self,
-            request: tonic::Request<super::UpdateFileRequest>,
-        ) -> std::result::Result<tonic::Response<super::FileResponse>, tonic::Status>;
-        async fn update_file_status(
-            &self,
-            request: tonic::Request<super::UpdateFileStatusRequest>,
-        ) -> std::result::Result<tonic::Response<super::FileResponse>, tonic::Status>;
-        async fn delete_file(
-            &self,
-            request: tonic::Request<super::DeleteFileRequest>,
-        ) -> std::result::Result<
-            tonic::Response<super::DeleteFileResponse>,
-            tonic::Status,
-        >;
-    }
-    #[derive(Debug)]
-    pub struct FileManagementServer<T> {
-        inner: Arc<T>,
-        accept_compression_encodings: EnabledCompressionEncodings,
-        send_compression_encodings: EnabledCompressionEncodings,
-        max_decoding_message_size: Option<usize>,
-        max_encoding_message_size: Option<usize>,
-    }
-    impl<T> FileManagementServer<T> {
-        pub fn new(inner: T) -> Self {
-            Self::from_arc(Arc::new(inner))
-        }
-        pub fn from_arc(inner: Arc<T>) -> Self {
-            Self {
-                inner,
-                accept_compression_encodings: Default::default(),
-                send_compression_encodings: Default::default(),
-                max_decoding_message_size: None,
-                max_encoding_message_size: None,
-            }
-        }
-        pub fn with_interceptor<F>(
-            inner: T,
-            interceptor: F,
-        ) -> InterceptedService<Self, F>
-        where
-            F: tonic::service::Interceptor,
-        {
-            InterceptedService::new(Self::new(inner), interceptor)
-        }
-        /// Enable decompressing requests with the given encoding.
-        #[must_use]
-        pub fn accept_compressed(mut self, encoding: CompressionEncoding) -> Self {
-            self.accept_compression_encodings.enable(encoding);
-            self
-        }
-        /// Compress responses with the given encoding, if the client supports it.
-        #[must_use]
-        pub fn send_compressed(mut self, encoding: CompressionEncoding) -> Self {
-            self.send_compression_encodings.enable(encoding);
-            self
-        }
-        /// Limits the maximum size of a decoded message.
-        ///
-        /// Default: `4MB`
-        #[must_use]
-        pub fn max_decoding_message_size(mut self, limit: usize) -> Self {
-            self.max_decoding_message_size = Some(limit);
-            self
-        }
-        /// Limits the maximum size of an encoded message.
-        ///
-        /// Default: `usize::MAX`
-        #[must_use]
-        pub fn max_encoding_message_size(mut self, limit: usize) -> Self {
-            self.max_encoding_message_size = Some(limit);
-            self
-        }
-    }
-    impl<T, B> tonic::codegen::Service<http::Request<B>> for FileManagementServer<T>
-    where
-        T: FileManagement,
-        B: Body + std::marker::Send + 'static,
-        B::Error: Into<StdError> + std::marker::Send + 'static,
-    {
-        type Response = http::Response<tonic::body::BoxBody>;
-        type Error = std::convert::Infallible;
-        type Future = BoxFuture<Self::Response, Self::Error>;
-        fn poll_ready(
-            &mut self,
-            _cx: &mut Context<'_>,
-        ) -> Poll<std::result::Result<(), Self::Error>> {
-            Poll::Ready(Ok(()))
-        }
-        fn call(&mut self, req: http::Request<B>) -> Self::Future {
-            match req.uri().path() {
-                "/file_management.FileManagement/CreateFile" => {
-                    #[allow(non_camel_case_types)]
-                    struct CreateFileSvc<T: FileManagement>(pub Arc<T>);
-                    impl<
-                        T: FileManagement,
-                    > tonic::server::UnaryService<super::CreateFileRequest>
-                    for CreateFileSvc<T> {
-                        type Response = super::FileResponse;
-                        type Future = BoxFuture<
-                            tonic::Response<Self::Response>,
-                            tonic::Status,
-                        >;
-                        fn call(
-                            &mut self,
-                            request: tonic::Request<super::CreateFileRequest>,
-                        ) -> Self::Future {
-                            let inner = Arc::clone(&self.0);
-                            let fut = async move {
-                                <T as FileManagement>::create_file(&inner, request).await
-                            };
-                            Box::pin(fut)
-                        }
-                    }
-                    let accept_compression_encodings = self.accept_compression_encodings;
-                    let send_compression_encodings = self.send_compression_encodings;
-                    let max_decoding_message_size = self.max_decoding_message_size;
-                    let max_encoding_message_size = self.max_encoding_message_size;
-                    let inner = self.inner.clone();
-                    let fut = async move {
-                        let method = CreateFileSvc(inner);
-                        let codec = tonic::codec::ProstCodec::default();
-                        let mut grpc = tonic::server::Grpc::new(codec)
-                            .apply_compression_config(
-                                accept_compression_encodings,
-                                send_compression_encodings,
-                            )
-                            .apply_max_message_size_config(
-                                max_decoding_message_size,
-                                max_encoding_message_size,
-                            );
-                        let res = grpc.unary(method, req).await;
-                        Ok(res)
-                    };
-                    Box::pin(fut)
-                }
-                "/file_management.FileManagement/UploadFile" => {
-                    #[allow(non_camel_case_types)]
-                    struct UploadFileSvc<T: FileManagement>(pub Arc<T>);
-                    impl<
-                        T: FileManagement,
-                    > tonic::server::ClientStreamingService<super::UploadFileRequest>
-                    for UploadFileSvc<T> {
-                        type Response = super::FileResponse;
-                        type Future = BoxFuture<
-                            tonic::Response<Self::Response>,
-                            tonic::Status,
-                        >;
-                        fn call(
-                            &mut self,
-                            request: tonic::Request<
-                                tonic::Streaming<super::UploadFileRequest>,
-                            >,
-                        ) -> Self::Future {
-                            let inner = Arc::clone(&self.0);
-                            let fut = async move {
-                                <T as FileManagement>::upload_file(&inner, request).await
-                            };
-                            Box::pin(fut)
-                        }
-                    }
-                    let accept_compression_encodings = self.accept_compression_encodings;
-                    let send_compression_encodings = self.send_compression_encodings;
-                    let max_decoding_message_size = self.max_decoding_message_size;
-                    let max_encoding_message_size = self.max_encoding_message_size;
-                    let inner = self.inner.clone();
-                    let fut = async move {
-                        let method = UploadFileSvc(inner);
-                        let codec = tonic::codec::ProstCodec::default();
-                        let mut grpc = tonic::server::Grpc::new(codec)
-                            .apply_compression_config(
-                                accept_compression_encodings,
-                                send_compression_encodings,
-                            )
-                            .apply_max_message_size_config(
-                                max_decoding_message_size,
-                                max_encoding_message_size,
-                            );
-                        let res = grpc.client_streaming(method, req).await;
-                        Ok(res)
-                    };
-                    Box::pin(fut)
-                }
-                "/file_management.FileManagement/GetFile" => {
-                    #[allow(non_camel_case_types)]
-                    struct GetFileSvc<T: FileManagement>(pub Arc<T>);
-                    impl<
-                        T: FileManagement,
-                    > tonic::server::UnaryService<super::GetFileRequest>
-                    for GetFileSvc<T> {
-                        type Response = super::FileResponse;
-                        type Future = BoxFuture<
-                            tonic::Response<Self::Response>,
-                            tonic::Status,
-                        >;
-                        fn call(
-                            &mut self,
-                            request: tonic::Request<super::GetFileRequest>,
-                        ) -> Self::Future {
-                            let inner = Arc::clone(&self.0);
-                            let fut = async move {
-                                <T as FileManagement>::get_file(&inner, request).await
-                            };
-                            Box::pin(fut)
-                        }
-                    }
-                    let accept_compression_encodings = self.accept_compression_encodings;
-                    let send_compression_encodings = self.send_compression_encodings;
-                    let max_decoding_message_size = self.max_decoding_message_size;
-                    let max_encoding_message_size = self.max_encoding_message_size;
-                    let inner = self.inner.clone();
-                    let fut = async move {
-                        let method = GetFileSvc(inner);
-                        let codec = tonic::codec::ProstCodec::default();
-                        let mut grpc = tonic::server::Grpc::new(codec)
-                            .apply_compression_config(
-                                accept_compression_encodings,
-                                send_compression_encodings,
-                            )
-                            .apply_max_message_size_config(
-                                max_decoding_message_size,
-                                max_encoding_message_size,
-                            );
-                        let res = grpc.unary(method, req).await;
-                        Ok(res)
-                    };
-                    Box::pin(fut)
-                }
-                "/file_management.FileManagement/ListFiles" => {
-                    #[allow(non_camel_case_types)]
-                    struct ListFilesSvc<T: FileManagement>(pub Arc<T>);
-                    impl<
-                        T: FileManagement,
-                    > tonic::server::UnaryService<super::ListFilesRequest>
-                    for ListFilesSvc<T> {
-                        type Response = super::ListFilesResponse;
-                        type Future = BoxFuture<
-                            tonic::Response<Self::Response>,
-                            tonic::Status,
-                        >;
-                        fn call(
-                            &mut self,
-                            request: tonic::Request<super::ListFilesRequest>,
-                        ) -> Self::Future {
-                            let inner = Arc::clone(&self.0);
-                            let fut = async move {
-                                <T as FileManagement>::list_files(&inner, request).await
-                            };
-                            Box::pin(fut)
-                        }
-                    }
-                    let accept_compression_encodings = self.accept_compression_encodings;
-                    let send_compression_encodings = self.send_compression_encodings;
-                    let max_decoding_message_size = self.max_decoding_message_size;
-                    let max_encoding_message_size = self.max_encoding_message_size;
-                    let inner = self.inner.clone();
-                    let fut = async move {
-                        let method = ListFilesSvc(inner);
-                        let codec = tonic::codec::ProstCodec::default();
-                        let mut grpc = tonic::server::Grpc::new(codec)
-                            .apply_compression_config(
-                                accept_compression_encodings,
-                                send_compression_encodings,
-                            )
-                            .apply_max_message_size_config(
-                                max_decoding_message_size,
-                                max_encoding_message_size,
-                            );
-                        let res = grpc.unary(method, req).await;
-                        Ok(res)
-                    };
-                    Box::pin(fut)
-                }
-                "/file_management.FileManagement/DownloadFile" => {
-                    #[allow(non_camel_case_types)]
-                    struct DownloadFileSvc<T: FileManagement>(pub Arc<T>);
-                    impl<
-                        T: FileManagement,
-                    > tonic::server::ServerStreamingService<super::DownloadFileRequest>
-                    for DownloadFileSvc<T> {
-                        type Response = super::FileChunk;
-                        type ResponseStream = T::DownloadFileStream;
-                        type Future = BoxFuture<
-                            tonic::Response<Self::ResponseStream>,
-                            tonic::Status,
-                        >;
-                        fn call(
-                            &mut self,
-                            request: tonic::Request<super::DownloadFileRequest>,
-                        ) -> Self::Future {
-                            let inner = Arc::clone(&self.0);
-                            let fut = async move {
-                                <T as FileManagement>::download_file(&inner, request).await
-                            };
-                            Box::pin(fut)
-                        }
-                    }
-                    let accept_compression_encodings = self.accept_compression_encodings;
-                    let send_compression_encodings = self.send_compression_encodings;
-                    let max_decoding_message_size = self.max_decoding_message_size;
-                    let max_encoding_message_size = self.max_encoding_message_size;
-                    let inner = self.inner.clone();
-                    let fut = async move {
-                        let method = DownloadFileSvc(inner);
-                        let codec = tonic::codec::ProstCodec::default();
-                        let mut grpc = tonic::server::Grpc::new(codec)
-                            .apply_compression_config(
-                                accept_compression_encodings,
-                                send_compression_encodings,
-                            )
-                            .apply_max_message_size_config(
-                                max_decoding_message_size,
-                                max_encoding_message_size,
-                            );
-                        let res = grpc.server_streaming(method, req).await;
-                        Ok(res)
-                    };
-                    Box::pin(fut)
-                }
-                "/file_management.FileManagement/UpdateFile" => {
-                    #[allow(non_camel_case_types)]
-                    struct UpdateFileSvc<T: FileManagement>(pub Arc<T>);
-                    impl<
-                        T: FileManagement,
-                    > tonic::server::UnaryService<super::UpdateFileRequest>
-                    for UpdateFileSvc<T> {
-                        type Response = super::FileResponse;
-                        type Future = BoxFuture<
-                            tonic::Response<Self::Response>,
-                            tonic::Status,
-                        >;
-                        fn call(
-                            &mut self,
-                            request: tonic::Request<super::UpdateFileRequest>,
-                        ) -> Self::Future {
-                            let inner = Arc::clone(&self.0);
-                            let fut = async move {
-                                <T as FileManagement>::update_file(&inner, request).await
-                            };
-                            Box::pin(fut)
-                        }
-                    }
-                    let accept_compression_encodings = self.accept_compression_encodings;
-                    let send_compression_encodings = self.send_compression_encodings;
-                    let max_decoding_message_size = self.max_decoding_message_size;
-                    let max_encoding_message_size = self.max_encoding_message_size;
-                    let inner = self.inner.clone();
-                    let fut = async move {
-                        let method = UpdateFileSvc(inner);
-                        let codec = tonic::codec::ProstCodec::default();
-                        let mut grpc = tonic::server::Grpc::new(codec)
-                            .apply_compression_config(
-                                accept_compression_encodings,
-                                send_compression_encodings,
-                            )
-                            .apply_max_message_size_config(
-                                max_decoding_message_size,
-                                max_encoding_message_size,
-                            );
-                        let res = grpc.unary(method, req).await;
-                        Ok(res)
-                    };
-                    Box::pin(fut)
-                }
-                "/file_management.FileManagement/UpdateFileStatus" => {
-                    #[allow(non_camel_case_types)]
-                    struct UpdateFileStatusSvc<T: FileManagement>(pub Arc<T>);
-                    impl<
-                        T: FileManagement,
-                    > tonic::server::UnaryService<super::UpdateFileStatusRequest>
-                    for UpdateFileStatusSvc<T> {
-                        type Response = super::FileResponse;
-                        type Future = BoxFuture<
-                            tonic::Response<Self::Response>,
-                            tonic::Status,
-                        >;
-                        fn call(
-                            &mut self,
-                            request: tonic::Request<super::UpdateFileStatusRequest>,
-                        ) -> Self::Future {
-                            let inner = Arc::clone(&self.0);
-                            let fut = async move {
-                                <T as FileManagement>::update_file_status(&inner, request)
-                                    .await
-                            };
-                            Box::pin(fut)
-                        }
-                    }
-                    let accept_compression_encodings = self.accept_compression_encodings;
-                    let send_compression_encodings = self.send_compression_encodings;
-                    let max_decoding_message_size = self.max_decoding_message_size;
-                    let max_encoding_message_size = self.max_encoding_message_size;
-                    let inner = self.inner.clone();
-                    let fut = async move {
-                        let method = UpdateFileStatusSvc(inner);
-                        let codec = tonic::codec::ProstCodec::default();
-                        let mut grpc = tonic::server::Grpc::new(codec)
-                            .apply_compression_config(
-                                accept_compression_encodings,
-                                send_compression_encodings,
-                            )
-                            .apply_max_message_size_config(
-                                max_decoding_message_size,
-                                max_encoding_message_size,
-                            );
-                        let res = grpc.unary(method, req).await;
-                        Ok(res)
-                    };
-                    Box::pin(fut)
-                }
-                "/file_management.FileManagement/DeleteFile" => {
-                    #[allow(non_camel_case_types)]
-                    struct DeleteFileSvc<T: FileManagement>(pub Arc<T>);
-                    impl<
-                        T: FileManagement,
-                    > tonic::server::UnaryService<super::DeleteFileRequest>
-                    for DeleteFileSvc<T> {
-                        type Response = super::DeleteFileResponse;
-                        type Future = BoxFuture<
-                            tonic::Response<Self::Response>,
-                            tonic::Status,
-                        >;
-                        fn call(
-                            &mut self,
-                            request: tonic::Request<super::DeleteFileRequest>,
-                        ) -> Self::Future {
-                            let inner = Arc::clone(&self.0);
-                            let fut = async move {
-                                <T as FileManagement>::delete_file(&inner, request).await
-                            };
-                            Box::pin(fut)
-                        }
-                    }
-                    let accept_compression_encodings = self.accept_compression_encodings;
-                    let send_compression_encodings = self.send_compression_encodings;
-                    let max_decoding_message_size = self.max_decoding_message_size;
-                    let max_encoding_message_size = self.max_encoding_message_size;
-                    let inner = self.inner.clone();
-                    let fut = async move {
-                        let method = DeleteFileSvc(inner);
-                        let codec = tonic::codec::ProstCodec::default();
-                        let mut grpc = tonic::server::Grpc::new(codec)
-                            .apply_compression_config(
-                                accept_compression_encodings,
-                                send_compression_encodings,
-                            )
-                            .apply_max_message_size_config(
-                                max_decoding_message_size,
-                                max_encoding_message_size,
-                            );
-                        let res = grpc.unary(method, req).await;
-                        Ok(res)
-                    };
-                    Box::pin(fut)
-                }
-                _ => {
-                    Box::pin(async move {
-                        let mut response = http::Response::new(empty_body());
-                        let headers = response.headers_mut();
-                        headers
-                            .insert(
-                                tonic::Status::GRPC_STATUS,
-                                (tonic::Code::Unimplemented as i32).into(),
-                            );
-                        headers
-                            .insert(
-                                http::header::CONTENT_TYPE,
-                                tonic::metadata::GRPC_CONTENT_TYPE,
-                            );
-                        Ok(response)
-                    })
-                }
-            }
-        }
-    }
-    impl<T> Clone for FileManagementServer<T> {
-        fn clone(&self) -> Self {
-            let inner = self.inner.clone();
-            Self {
-                inner,
-                accept_compression_encodings: self.accept_compression_encodings,
-                send_compression_encodings: self.send_compression_encodings,
-                max_decoding_message_size: self.max_decoding_message_size,
-                max_encoding_message_size: self.max_encoding_message_size,
-            }
-        }
-    }
-    /// Generated gRPC service name
-    pub const SERVICE_NAME: &str = "file_management.FileManagement";
-    impl<T> tonic::server::NamedService for FileManagementServer<T> {
-        const NAME: &'static str = SERVICE_NAME;
-    }
-}
