--- src/models/_proto/mod.rs
@@ -1,5 +1,0 @@
-pub mod auth;
-pub mod employee_management;
-pub mod file_management;
-pub mod health_check;
-pub mod storage;
