--- src/models/mod.rs
@@ -1,3 +1,0 @@
-pub mod _entities;
-pub mod _proto;
-pub mod user;
